+++ Group/src/Driver.java
@@ -1,5 +1,8 @@
+import java.io.File;
+import java.io.IOException;
+
 public class Driver {
-  public static void main(String[] args) {
-     // NOTE: The actual driver program will run tests on the above objects
-  }
+    public static void main(String[] args) throws IOException {
+        CompressFile.compress(new File("input.txt"), new File("output.txt"));
+    }
 }
